+++ src/test/java/org/aldinrizvo/qamp/API_Homework2/Task2/tests/LoginWithEmptyPasswordFieldTest.java
@@ -1,4 +1,4 @@
-package org.aldinrizvo.qamp.API_Homework2.Task2.Tests;
+package org.aldinrizvo.qamp.API_Homework2.Task2.tests;
 
 import io.restassured.response.Response;
 import org.aldinrizvo.qamp.API_Homework2.Task2.steps.HttpRequests;
@@ -8,16 +8,16 @@ import org.testng.annotations.Test;
 
 import java.util.logging.Logger;
 
-public class LoginWithEmptyEmailFieldTest {
-    private static final Logger LOGGER = Logger.getLogger(LoginWithEmptyEmailFieldTest.class.getName());
+public class LoginWithEmptyPasswordFieldTest {
+    private static final Logger LOGGER = Logger.getLogger(LoginWithEmptyPasswordFieldTest.class.getName());
 
     @Test(
             priority = 1,
-            description = "This Tests verifies that user can't sign in with invalid email.."
+            description = "This tests verifies that user can't sign in with invalid email."
     )
     public void testLoginPageNegative() {
         final AuthenticationRequest authenticationRequest =
-                new AuthenticationRequest("", GlobalValues.PASSWORD);
+                new AuthenticationRequest(GlobalValues.EMAIL, "");
 
         final Response response = HttpRequests.sendWwwFormUrlEncodedPostRequest(
                 GlobalValues.AUTH,
